+++ integrations/node/oicana-node/src/Template.ts
@@ -11,19 +11,20 @@ import {
   getSource,
   getWarnings,
   CompilationMode as NativeCompilationMode,
-  inputs as nativeInputs,
+  manifest as nativeManifest,
   registerTemplate,
   registerTemplateAsync,
   removeDocument,
   removeWorld,
   setValidateInputs,
 } from '@oicana/node-native';
+import type { BlobInput } from './BlobInput.js';
 import { CompilationMode } from './CompilationMode.js';
 import { CompiledDocument } from './CompiledDocument.js';
 import { type ExportFormat, Pdf, Png, Svg } from './ExportFormat.js';
 import type { ExportOnceResult } from './ExportOnceResult.js';
-import type { BlobWithMetadata } from './inputs/index.js';
 import { type PageRange, serializePageRange } from './PageRange.js';
+import type { TemplateManifest } from './TemplateManifest.js';
 import type { ZipLimits } from './ZipLimits.js';
 
 /**
@@ -76,7 +77,7 @@ export class Template implements Disposable {
   public constructor(
     template: Uint8Array,
     jsonInputs: Map<string, string>,
-    blobInputs: Map<string, BlobWithMetadata>,
+    blobInputs: Map<string, BlobInput>,
   );
 
   /**
@@ -90,7 +91,7 @@ export class Template implements Disposable {
   public constructor(
     template: Uint8Array,
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     limits?: ZipLimits,
   );
@@ -98,7 +99,7 @@ export class Template implements Disposable {
   public constructor(
     template: Uint8Array | CompletedRegistration,
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     limits?: ZipLimits,
   ) {
@@ -120,9 +121,7 @@ export class Template implements Disposable {
       this.template,
       template,
       Object.fromEntries(jsonInputs ?? new Map<string, string>()),
-      Template.convertBlobWithMetadata(
-        blobInputs ?? new Map<string, BlobWithMetadata>(),
-      ),
+      Template.convertBlobInputs(blobInputs ?? new Map<string, BlobInput>()),
       Template.mapCompilationMode(
         compilationOptions ?? CompilationMode.Development,
       ),
@@ -147,7 +146,7 @@ export class Template implements Disposable {
   public static async create(
     template: Uint8Array,
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     limits?: ZipLimits,
   ): Promise<Template> {
@@ -157,9 +156,7 @@ export class Template implements Disposable {
       templateId,
       template,
       Object.fromEntries(jsonInputs ?? new Map<string, string>()),
-      Template.convertBlobWithMetadata(
-        blobInputs ?? new Map<string, BlobWithMetadata>(),
-      ),
+      Template.convertBlobInputs(blobInputs ?? new Map<string, BlobInput>()),
       Template.mapCompilationMode(
         compilationOptions ?? CompilationMode.Development,
       ),
@@ -193,7 +190,7 @@ export class Template implements Disposable {
   public static exportOnce(
     template: Uint8Array,
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     exportOptions?: ExportFormat,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
@@ -202,9 +199,7 @@ export class Template implements Disposable {
     const result = exportTemplateOnce(
       template,
       Object.fromEntries(jsonInputs ?? new Map<string, string>()),
-      Template.convertBlobWithMetadata(
-        blobInputs ?? new Map<string, BlobWithMetadata>(),
-      ),
+      Template.convertBlobInputs(blobInputs ?? new Map<string, BlobInput>()),
       Template.mapCompilationMode(
         compilationOptions ?? CompilationMode.Production,
       ),
@@ -231,7 +226,7 @@ export class Template implements Disposable {
   public static async exportOnceAsync(
     template: Uint8Array,
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     exportOptions?: ExportFormat,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
@@ -240,9 +235,7 @@ export class Template implements Disposable {
     const result = await exportTemplateOnceAsync(
       template,
       Object.fromEntries(jsonInputs ?? new Map<string, string>()),
-      Template.convertBlobWithMetadata(
-        blobInputs ?? new Map<string, BlobWithMetadata>(),
-      ),
+      Template.convertBlobInputs(blobInputs ?? new Map<string, BlobInput>()),
       Template.mapCompilationMode(
         compilationOptions ?? CompilationMode.Production,
       ),
@@ -267,7 +260,7 @@ export class Template implements Disposable {
    */
   public export(
     jsonInputs: Map<string, string>,
-    blobInputs: Map<string, BlobWithMetadata>,
+    blobInputs: Map<string, BlobInput>,
   ): Uint8Array;
 
   /**
@@ -278,7 +271,7 @@ export class Template implements Disposable {
    */
   public export(
     jsonInputs: Map<string, string>,
-    blobInputs: Map<string, BlobWithMetadata>,
+    blobInputs: Map<string, BlobInput>,
     exportOptions: ExportFormat,
   ): Uint8Array;
 
@@ -291,7 +284,7 @@ export class Template implements Disposable {
    */
   public export(
     jsonInputs: Map<string, string>,
-    blobInputs: Map<string, BlobWithMetadata>,
+    blobInputs: Map<string, BlobInput>,
     exportOptions: ExportFormat,
     compilationOptions: CompilationMode,
   ): Uint8Array;
@@ -306,7 +299,7 @@ export class Template implements Disposable {
    */
   public export(
     jsonInputs: Map<string, string>,
-    blobInputs: Map<string, BlobWithMetadata>,
+    blobInputs: Map<string, BlobInput>,
     exportOptions: ExportFormat,
     compilationOptions: CompilationMode,
     pages: PageRange,
@@ -327,7 +320,7 @@ export class Template implements Disposable {
    */
   public export(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     exportOptions?: ExportFormat,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
@@ -355,7 +348,7 @@ export class Template implements Disposable {
    */
   public exportAsync(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     exportOptions?: ExportFormat,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
@@ -380,7 +373,7 @@ export class Template implements Disposable {
    */
   public exportPdf(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
   ): Uint8Array {
@@ -405,7 +398,7 @@ export class Template implements Disposable {
    */
   public exportPdfAsync(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
   ): Promise<Uint8Array> {
@@ -430,7 +423,7 @@ export class Template implements Disposable {
    */
   public exportPng(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pixelsPerPt = 1.0,
     pages?: PageRange,
@@ -457,7 +450,7 @@ export class Template implements Disposable {
    */
   public exportPngAsync(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pixelsPerPt = 1.0,
     pages?: PageRange,
@@ -481,7 +474,7 @@ export class Template implements Disposable {
    */
   public exportSvg(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
   ): Uint8Array {
@@ -505,7 +498,7 @@ export class Template implements Disposable {
    */
   public exportSvgAsync(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
   ): Promise<Uint8Array> {
@@ -521,7 +514,7 @@ export class Template implements Disposable {
   private exportWith(
     format: ExportFormat,
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
   ): Uint8Array {
@@ -544,7 +537,7 @@ export class Template implements Disposable {
   private async exportWithAsync(
     format: ExportFormat,
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
     pages?: PageRange,
   ): Promise<Uint8Array> {
@@ -577,7 +570,7 @@ export class Template implements Disposable {
    */
   public compile(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
   ): CompiledDocument {
     const documentId = this.compileToDocumentId(
@@ -602,7 +595,7 @@ export class Template implements Disposable {
    */
   public async compileAsync(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
   ): Promise<CompiledDocument> {
     const documentId = await this.compileToDocumentIdAsync(
@@ -615,15 +608,13 @@ export class Template implements Disposable {
 
   private compileToDocumentId(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
   ): string {
     const documentId = compileTemplate(
       this.template,
       Object.fromEntries(jsonInputs ?? new Map<string, string>()),
-      Template.convertBlobWithMetadata(
-        blobInputs ?? new Map<string, BlobWithMetadata>(),
-      ),
+      Template.convertBlobInputs(blobInputs ?? new Map<string, BlobInput>()),
       Template.mapCompilationMode(
         compilationOptions ?? CompilationMode.Production,
       ),
@@ -634,15 +625,13 @@ export class Template implements Disposable {
 
   private async compileToDocumentIdAsync(
     jsonInputs?: Map<string, string>,
-    blobInputs?: Map<string, BlobWithMetadata>,
+    blobInputs?: Map<string, BlobInput>,
     compilationOptions?: CompilationMode,
   ): Promise<string> {
     const documentId = await compileTemplateAsync(
       this.template,
       Object.fromEntries(jsonInputs ?? new Map<string, string>()),
-      Template.convertBlobWithMetadata(
-        blobInputs ?? new Map<string, BlobWithMetadata>(),
-      ),
+      Template.convertBlobInputs(blobInputs ?? new Map<string, BlobInput>()),
       Template.mapCompilationMode(
         compilationOptions ?? CompilationMode.Production,
       ),
@@ -660,10 +649,13 @@ export class Template implements Disposable {
   }
 
   /**
-   * The template's input definitions from its manifest, serialized as JSON.
+   * The template's manifest.
+   *
+   * The Typst package section und its Oicana
+   * configuration with the input definitions.
    */
-  public inputs(): string {
-    return nativeInputs(this.template);
+  public manifest(): TemplateManifest {
+    return JSON.parse(nativeManifest(this.template));
   }
 
   /**
@@ -708,14 +700,17 @@ export class Template implements Disposable {
     this.dispose();
   }
 
-  private static convertBlobWithMetadata(
-    blobInputs: Map<string, BlobWithMetadata>,
+  private static convertBlobInputs(
+    blobInputs: Map<string, BlobInput>,
   ): Record<string, BlobWithMetadataNative> {
     return Object.fromEntries(
       Array.from(blobInputs.entries(), ([key, value]) => {
         const nativeValue = {
-          bytes: value.bytes,
-          meta: value.meta === undefined ? '{}' : JSON.stringify(value.meta),
+          bytes: value.data,
+          meta:
+            value.metadata === undefined
+              ? '{}'
+              : JSON.stringify(value.metadata),
         };
         return [key, nativeValue];
       }),